hash/maphash: move purego-only helper functions to build tagged file
Hash.float64 and btoi helper functions are used only in the purego
version. Move them to the build tagged file.

Reviewed-on: https://go-review.googlesource.com/c/go/+/675158
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
